+++ inst/doc/bsims01-intro.R
@@ -1,0 +1,9 @@
+## ----setup,include=FALSE------------------------------------------------------
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+par(mar = c(1, 1, 1, 1))
+set.seed(429)
+suppressPackageStartupMessages(library(bSims))
+