+++ inst/shiny/bsimsH.R
@@ -2,6 +2,7 @@ library(shiny)
 library(detect)
 library(bSims)
 
+MAXDIS <- 20
 EXTENT <- 10
 DURATION <- 10
 TINT <- list(
@@ -15,6 +16,7 @@ TINT <- list(
 RINT <- list(
   "0-50-100-Inf m"=c(0.5, 1, Inf),
   "0-Inf m"=c(Inf),
+  "0-50-Inf m"=c(0.5, Inf),
   "0-50-100-150-Inf m"=c(0.5, 1, 1.5, Inf),
   "0-50-100-150-200-Inf m"=c(0.5, 1, 1.5, 2, Inf),
   "0-50-100 m"=c(0.5, 1),
@@ -100,7 +102,8 @@ ui <- navbarPage("bSims (H)",
       sliderInput("phim", "Movement rate", 0, 10, 1, 0.1),
       sliderInput("SDm", "Movement SD", 0, 1, 0, 0.05),
       checkboxInput("overlap", "Territory overlap allowed", TRUE),
-      checkboxInput("show_tess", "Show tessellation", FALSE)
+      checkboxInput("show_tess", "Show tessellation", FALSE),
+      checkboxInput("init_loc", "Initial location", FALSE)
     )
   ),
   tabPanel("Detect",
@@ -109,8 +112,9 @@ ui <- navbarPage("bSims (H)",
       plotOutput(outputId = "plot_dfun")
     ),
     column(6,
-      sliderInput("tau", "Detection parameter (tau)", 0, 5, 1, 0.25),
-      sliderInput("bpar", "Hazard rate parameter (b)", 0, 5, 1, 0.5),
+      sliderInput("tauV", "Detection parameter (tau), vocalizations", 0, MAXDIS, 1, MAXDIS/200),
+      sliderInput("tauM", "Detection parameter (tau), movement", 0, MAXDIS, 1, MAXDIS/200),
+      sliderInput("bpar", "Hazard rate parameter (b), vocalizations", 0, MAXDIS, 1, MAXDIS/200),
       radioButtons("dfun", "Distance function",
         c("Half Normal"="halfnormal",
           "Negative Exponential"="negexp",
@@ -118,7 +122,8 @@ ui <- navbarPage("bSims (H)",
       radioButtons("event", "Event type",
         c("Vocalization"="vocal",
           "Movement"="move",
-          "Both"="both"))
+          "Both"="both")),
+      sliderInput("dir_sens", "Anisotropy in tau due to direction", 0, 2, 1, 0.1)
     )
   ),
   tabPanel("Transcribe",
@@ -130,6 +135,7 @@ ui <- navbarPage("bSims (H)",
         selectInput("tint", "Time intervals", names(TINT)),
         selectInput("rint", "Distance intervals", names(RINT)),
         sliderInput("derr", "Distance error", 0, 1, 0, 0.1),
+        sliderInput("dbias", "Distance bias", 0, 2, 1, 0.1),
         radioButtons("condition", "Condition",
           c("1st event"="event1",
             "1st detection"="det1",
@@ -159,6 +165,12 @@ ui <- navbarPage("bSims (H)",
       verbatimTextOutput("settings"),
       uiOutput("clip")
     )
+  ),
+  tabPanel("Documentation",
+    column(12,
+      tags$iframe(src="https://psolymos.github.io/bSims/",
+        height=600, width="100%", frameBorder=0)
+    )
   )
 )
 
@@ -166,7 +178,7 @@ server <- function(input, output) {
   observeEvent(input$seed, {
     rv$seed <- rv$seed + 1
   })
-  dis <- seq(0, 10, 0.01)
+  dis <- seq(0, MAXDIS, MAXDIS/200)
   l <- reactive({
     set.seed(rv$seed)
     bsims_init(extent = EXTENT)
@@ -197,7 +209,8 @@ server <- function(input, output) {
       move_rate = input$phim,
       movement = input$SDm,
       mixture = c(input$mix, 1-input$mix),
-      allow_overlap = input$overlap)
+      allow_overlap = input$overlap,
+      initial_location = input$init_loc)
   })
   dfun <- reactive({
     switch(input$dfun,
@@ -209,10 +222,11 @@ server <- function(input, output) {
   o <- reactive({
     bsims_detect(b(),
       xy = c(0, 0),
-      tau = input$tau,
+      tau = c(input$tauV, input$tauM),
       dist_fun = dfun(),
-#      repel = input$repel,
-      event_type = input$event)
+      event_type = input$event,
+      direction=input$dir_sens != 1,
+      sensitivity=input$dir_sens)
   })
   m <- reactive({
     pr <- if (!input$oucount)
@@ -221,6 +235,7 @@ server <- function(input, output) {
       tint = TINT[[input$tint]],
       rint = RINT[[input$rint]],
       error = input$derr,
+      bias = input$dbias,
       condition = input$condition,
       event_type = input$event,
       perception = pr
@@ -235,21 +250,34 @@ server <- function(input, output) {
     Ydis <- matrix(rowSums(REM), 1)
     Ddis <- matrix(RINT[[input$rint]], 1)
     if (length(TINT[[input$tint]]) > 1 && sum(REM) > 0) {
-      Mrem <- cmulti.fit(Ydur, Ddur, type="rem")
-      phi <- exp(Mrem$coef)
-      p <- 1-exp(-MaxDur*phi)
+      Mrem <- try(cmulti.fit(Ydur, Ddur, type="rem"))
+      if (!inherits(Mrem, "try-error")) {
+        phi <- exp(Mrem$coef)
+        p <- 1-exp(-MaxDur*phi)
+      } else {
+        Mrem <- NULL
+        phi <- NA
+        p <- NA
+      }
     } else {
       Mrem <- NULL
       phi <- NA
       p <- NA
     }
     if (length(RINT[[input$rint]]) > 1 && sum(REM) > 0) {
-      Mdis <- cmulti.fit(Ydis, Ddis, type="dis")
-      tau <- exp(Mdis$coef)
-      q <- if (is.infinite(MaxDis))
-        1 else (tau^2/MaxDis^2) * (1-exp(-(MaxDis/tau)^2))
-      A <- if (is.infinite(MaxDis))
-        pi * tau^2 else pi * MaxDis^2
+      Mdis <- try(cmulti.fit(Ydis, Ddis, type="dis"))
+      if (!inherits(Mdis, "try-error")) {
+        tau <- exp(Mdis$coef)
+        q <- if (is.infinite(MaxDis))
+          1 else (tau^2/MaxDis^2) * (1-exp(-(MaxDis/tau)^2))
+        A <- if (is.infinite(MaxDis))
+          pi * tau^2 else pi * MaxDis^2
+      } else {
+        Mdis <- NULL
+        tau <- NA
+        q <- NA
+        A <- NA
+      }
     } else {
       Mdis <- NULL
       tau <- NA
@@ -286,15 +314,18 @@ server <- function(input, output) {
     ",\n  movement = ", input$SDm,
     ",\n  mixture = ", xc(c(input$mix, 1-input$mix)),
     ",\n  allow_overlap = ", input$overlap,
-    ",\n  tau = ", input$tau,
+    ",\n  initial_location = ", input$init_loc,
+    ",\n  tau = c(", input$tauV, ", ", input$tauM, ")",
     ",\n  dist_fun = ", paste0(deparse(dfun()), collapse=''),
     ",\n  xy = c(0, 0)",
+    ",\n  direction = ", input$dir_sens != 1,
+    ",\n  sensitivity = ", input$dir_sens,
     ",\n  event_type = ", xq(input$event),
     ",\n  tint = ", xc(TINT[[input$tint]]),
     ",\n  rint = ", xc(RINT[[input$rint]]),
     ",\n  error = ", input$derr,
+    ",\n  bias = ", input$dbias,
     ",\n  condition = ", xq(input$condition),
-    ",\n  event_type = ", xq(input$event),
     ",\n  perception = ", pr,
     ")", collapse="")
   })
@@ -315,9 +346,10 @@ server <- function(input, output) {
   })
   output$plot_ani <- renderPlot({
     op <- par(mar=c(0,0,0,0))
-    plot(b())
+    plot(b(), event_type=input$event)
     if (input$show_tess && !is.null(b()$tess))
-      plot(b()$tess, TRUE, "tess", "none", col="grey", lty=1)
+      plot(b()$tess, add=TRUE, wlines="tess",
+        showpoints=FALSE, cmpnt_col="grey", cmpnt_lty=1)
     par(op)
   })
   output$plot_det <- renderPlot({
@@ -328,8 +360,11 @@ server <- function(input, output) {
     par(op)
   })
   output$plot_dfun <- renderPlot({
-    plot(dis, dfun()(dis, input$tau), type="l", col=4,
+    plot(dis, dfun()(dis, input$tauV), type="l", col="black",
       ylim=c(0,1), xlab="Distance", ylab="P(detection)")
+    lines(dis, dfun()(dis, input$tauM), col="purple")
+    legend("top", horiz=TRUE, bty="n", lty=1, col=c("black","purple"),
+      legend=c("vocalization", "movement"))
   })
   output$plot_tra <- renderPlot({
     op <- par(mar=c(0,0,0,0))
@@ -348,8 +383,16 @@ server <- function(input, output) {
     op <- par(mfrow=c(1,3))
     barplot(c(True=input$phi1, Estimate=v$phi),
       col=col, main=expression(phi))
-    barplot(c(True=input$tau, Estimate=v$tau),
-      col=col, main=expression(tau))
+    Taus <- switch(input$event,
+      "vocal"=c(True=input$tauV, Estimate=v$tau),
+      "move"=c(True=input$tauM, Estimate=v$tau),
+      "both"=c(TrueV=input$tauV, TrueM=input$tauM, Estimate=v$tau))
+    TauCol <- switch(input$event,
+      "vocal"=col[c(1,2)],
+      "move"=col[c(1,2)],
+      "both"=col[c(1,1,2)])
+    barplot(Taus,
+      col=TauCol, main=expression(tau))
     barplot(c(True=input$D, Estimate=v$D),
       col=col, main=expression(D))
     par(op)